3.0
FINE

+++ AppStar/src/control/RicercaStelleInFilamentoController.java
@@ -18,26 +18,26 @@ public class RicercaStelleInFilamentoController {
         ricercaStelleInFilamentoGUI.istanziaRicercaStelleInFilamentoGUIFXML(e);
     }
 
-    public ArrayList<Float> cercaInFilamento(ObservableList<Stella> stella, TableView tableView, TableColumn id,
+    public ArrayList<Double> cercaInFilamento(ObservableList<Stella> stella, TableView tableView, TableColumn id,
                                              TableColumn nameStar, TableColumn glon, TableColumn glat,
                                              TableColumn flux, TableColumn type, int idFil, String satellite,
                                              int pagina){
         ArrayList<Integer> arrayList;
-        ArrayList<Float> arrayPercentuali = new ArrayList<>(4);
+        ArrayList<Double> arrayPercentuali = new ArrayList<>(4);
 
         UtenteRegistrato utente = UtenteConnesso.getInstance(null, null);
         arrayList = utente.cercaInFilamento(stella, tableView, id, nameStar, glon, glat, flux, type, idFil, satellite,
                 pagina);
 
-        float prestellar = Float.parseFloat(arrayList.get(0).toString());
-        float protostellar = Float.parseFloat(arrayList.get(1).toString());
-        float unbound = Float.parseFloat(arrayList.get(2).toString());
+        double prestellar = Double.parseDouble(arrayList.get(0).toString());
+        double protostellar = Double.parseDouble(arrayList.get(1).toString());
+        double unbound = Double.parseDouble(arrayList.get(2).toString());
 
-         float totale = unbound + prestellar + protostellar;
-         float unboundPerc = (unbound/totale)*100;
-         float prestellarPerc = (prestellar/totale)*100;
-         float protostellarPerc = (protostellar/totale)*100;
-         arrayPercentuali.add(0, totale); arrayPercentuali.add(1, unboundPerc);
+        double totale = unbound + prestellar + protostellar;
+        double unboundPerc = (unbound/totale)*100;
+        double prestellarPerc = (prestellar/totale)*100;
+        double protostellarPerc = (protostellar/totale)*100;
+        arrayPercentuali.add(0, totale); arrayPercentuali.add(1, unboundPerc);
         arrayPercentuali.add(2, prestellarPerc); arrayPercentuali.add(3, protostellarPerc);
             //arrayPercentuali contiene [totale stelle trovate, percentuale unbound, percentule prestellar,
              // percentuale protostellar].
